UPSTREAM: netfilter: xt_bpf: support ebpf
Add support for attaching an eBPF object by file descriptor.

The iptables binary can be called with a path to an elf object or a
pinned bpf object. Also pass the mode and path to the kernel to be
able to return it later for iptables dump and save.

Signed-off-by: Willem de Bruijn <willemb@google.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>
Signed-off-by: Chenbo Feng <fengc@google.com>
(cherry picked from commit 2c16d60332643e90d4fa244f4a706c454b8c7569)

Change-Id: I31b8831a7ffd7c44985ee906ff194c1d934dafbe
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>
Signed-off-by: Omar Mohamed <omarcoptan9@gmail.com>