kernel: power: wakelock: Add poor hack for sensor_ind wakelock
Add a very poor hack that converts the sensor_ind wakelock to
a timed one (with a 5 seconds timeout) on-the-fly.
This is unfortunately necessary, as it is the only solution for
a battery drainage issue, for broken DSP firmwares on old
SoCs combined with proprietary sensor binaries which aren't
accounting for that (anymore).

This change will somewhat retain the functionality that these
sensor binaries are aiming to give on a system featuring an
Android userspace.

Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>
Signed-off-by: Omar Mohamed <omarcoptan9@gmail.com>