prlmk: Handle a critical lowmemory situation
Now that we've set the appropriate notifiers in is_low_mem(),
we use that to tell us when the memory situation is critically
low, and accordingly call pagefault_out_of_memory(), which will
kill tasks that consumes the most amount of memory (i.e, memory hogs).

pagefault_out_of_memory() is defined in mm/oom_kill.c

This is a temporary fix until a better alternative can be found.

Signed-off-by: darkhz <kmachanwenw@gmail.com>
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>
Signed-off-by: Omar Mohamed <omarcoptan9@gmail.com>